Various improvements according to comments
