+++ VersionControlUI/src/items/VDiffComparisonPair.h
@@ -26,18 +26,38 @@
 
 #pragma once
 
-#include "versioncontrolui_api.h"
+#include "../versioncontrolui_api.h"
 
+#include "VisualizationBase/src/items/ItemWithNode.h"
+#include "VisualizationBase/src/declarative/DeclarativeItem.h"
 #include "VisualizationBase/src/declarative/DeclarativeItemBaseStyle.h"
-#include "VisualizationBase/src/items/TextStyle.h"
+#include "VDiffComparisonPairStyle.h"
 
-namespace VERSIONCONTROLUI_API
+#include "VisualizationBase/src/items/Item.h"
+#include "VisualizationBase/src/items/VText.h"
+
+#include "../nodes/DiffComparisonPair.h"
+
+namespace VersionControlUI
 {
 
-class VERSIONCONTROLUI_API VDiffComparisonPairStyle : public Super<Visualization::DeclarativeItemBaseStyle>
+class DiffComparisonPair;
+
+class VERSIONCONTROLUI_API VDiffComparisonPair : public Super<Visualization::ItemWithNode<VDiffComparisonPair,
+		Visualization::DeclarativeItem<VDiffComparisonPair>, DiffComparisonPair>>
 {
+	ITEM_COMMON(VDiffComparisonPair)
+
 	public:
-		Property<Visualization::TextStyle> oldVersionObjectPath{this, "oldVersionObjectPath"};
-		Property<Visualization::TextStyle> newVersionObjectPath{this, "newVersionObjectPath"};
+		VDiffComparisonPair(Visualization::Item* parent, NodeType* node, const StyleType* style = itemStyles().get());
+		static void initializeForms();
+		virtual int determineForm() override;
+
+	private:
+		Visualization::Item* oldVersionNode_{};
+		Visualization::Item* newVersionNode_{};
+		Visualization::VText* oldVersionObjectPath_{};
+		Visualization::VText* newVersionObjectPath_{};
 };
+
 }
